server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
google-generativeai==0.8.4
gunicorn==23.0.0
httpx==0.27.2
python-dotenv==1.0.1
youtube-search-python==1.6.6
youtube-transcript-api==1.2.4
